chore:tada: create home page
