+++ src/App.css
@@ -1,38 +1,78 @@
 .App {
   text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
+  background-color: rgba(206, 60, 66, 0.3);
   min-height: 100vh;
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  font-size: calc(10px + 2vmin);
+}
+
+.App-header {
+  display: flex;
+  align-items: center;
+  justify-content: center;
   color: white;
 }
 
-.App-link {
-  color: #61dafb;
+.App-header img {
+  max-width: 18%;
+  object-fit: contain;
 }
 
-@keyframes App-logo-spin {
-  from {
+.App-title {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 32px;
+  font-weight: bold;
+}
+
+.App-button {
+  position: relative;
+  border-radius: 20px;
+  cursor: pointer;
+  padding: 4px 20px;
+  font-size: 28px;
+  background: #009752;
+  color: #fff;
+  -webkit-box-shadow: 0 5px 0 #57a3b8;
+  box-shadow: 0 5px 0 #57a3b8;
+}
+
+.App-button:hover {
+  -webkit-transform: translate(0, 3px);
+  transform: translate(0, 3px);
+  -webkit-box-shadow: 0 2px 0 #57a3b8;
+  box-shadow: 0 2px 0 #57a3b8;
+}
+
+.Loading-container {
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(206, 60, 66, 0.3);
+}
+
+.Loading {
+  box-sizing: border-box;
+  width: 1.2rem;
+  height: 1.2rem;
+  border-radius: 50%;
+  box-shadow: 0 -2rem 0 #eee, 1.4rem -1.4rem 0 #ddd, 2rem 0 0 #ccc,
+    1.4rem 1.4rem 0 rgba(0, 150, 82, 0.2), 0 2rem 0 rgba(0, 150, 82, 0.4),
+    -1.4rem 1.4rem 0 rgba(0, 150, 82, 0.6), -2rem 0 0 rgba(0, 150, 82, 0.8),
+    -1.4rem -1.4rem 0 #009752;
+  animation: loading 1s steps(8) 0s infinite;
+}
+
+@keyframes loading {
+  0% {
     transform: rotate(0deg);
   }
-  to {
+  100% {
     transform: rotate(360deg);
   }
 }
