+++ src/components/Loading.tsx
@@ -1,0 +1,7 @@
+export const Loading = () => {
+  return (
+    <div className="Loading-container">
+      <div className="Loading"></div>
+    </div>
+  );
+};
