+++ src/types/Status.ts
@@ -1,0 +1,5 @@
+export enum Status {
+  FAR = "FAR",
+  NEAR = "NEAR",
+  HIT = "HIT",
+}
